pass down the TestFramework's arguments to SBTRunner
This gives the TestFramework's arguments to SBTRunner in
scala-partest, so that it can support the same input as the command line
version.

This assumes the necessary changes were done on the scala-partest side.

With both this change and the one in scala-partest, "sbt-partest" is
just as flexible as "ant-partest", using `testOnly`. For example:

```
sbt> testOnly -- files/pos/t6745.scala --update-check
sbt> testOnly -- --show-diff pos scalacheck
```
(the leading `--` is necessary for now, sadly...)